+++ src/main/java/com/codiapp/codi/domain/login/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.codiapp.codi.domain.login.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long>{
     Optional<User> findByEmail(String email);    
 }
