+++ src/main/java/com/codiapp/codi/domain/team/dto/request/TeamUpdateRequestDTO.java
@@ -2,7 +2,16 @@ package com.codiapp.codi.domain.team.dto.request;
 
 import java.util.List;
 
-public record TeamUpdateRequestDTO (
-	String name,
-	List<String> memberEmails
-) {}
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TeamUpdateRequestDTO {
+	private String name;
+	private List<String> memberEmails;
+}
