--- src/main/java/com/codiapp/codi/domain/team/entity/User.java
@@ -1,34 +1,0 @@
-package com.codiapp.codi.domain.team.entity;
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "users")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "user_seq")
-    @SequenceGenerator(name = "user_seq", sequenceName = "USER_SEQ", allocationSize = 1)
-    private Long id;
-
-    @Column(nullable = false, unique = true, length = 50)
-    private String email;
-
-    @Column(nullable = false, length = 100)
-    private String password;
-
-    @Column(name = "username", nullable = false, length = 20)
-    private String userName;
-
-    @Column(name = "birth_date", nullable = false)
-    private LocalDate birthDate;
-
-    @Column(nullable = false)
-    private String status = "ACTIVE";
-}
-
