--- src/main/java/com/codiapp/codi/domain/team/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.codiapp.codi.domain.team.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.codiapp.codi.domain.team.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
