+++ src/main/java/com/codiapp/codi/global/config/WebConfig.java
@@ -7,14 +7,14 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") // 모든 API 경로 허용
-                .allowedOrigins("http://localhost:3000") // React 개발서버 출처 허용
-                .allowedMethods("GET", "POST", "PATCH", "PUT", "DELETE", "OPTIONS")
-                .allowedHeaders("*")
-                .allowCredentials(true); 
-    }
+//    @Override
+//    public void addCorsMappings(CorsRegistry registry) {
+//        registry.addMapping("/**") // 모든 API 경로 허용
+//                .allowedOrigins("http://localhost:3000") // React 개발서버 출처 허용
+//                .allowedMethods("GET", "POST", "PATCH", "PUT", "DELETE", "OPTIONS")
+//                .allowedHeaders("*")
+//                .allowCredentials(true); 
+//    }
 
     // 게시글 이미지
     @Override
